+++ tests/allocation.test.jsx
@@ -207,6 +207,6 @@ describe('the screen', () => {
       teams: TEAMS,
     })
     await setup()
-    expect(await screen.findByText(/haven't been given the Pitch Manager job/i)).toBeInTheDocument()
+    expect(await screen.findByText(/Pitch Management hasn't been added to your account/i)).toBeInTheDocument()
   })
 })